[1422] Move context menu-related contributions to syson-components
Bug: https://github.com/eclipse-syson/syson/issues/1422
